Move duplicate-looking test code to datasets
The "duplicate code" is actually a test data set
represented in PHP
