+++ tests/src/Middleware/Cli/OptionsTest.php
@@ -159,48 +159,7 @@ class OptionsTest extends PHPUnit_Framework_TestCase
      */
     public function provideOptionConfigsForStringTest()
     {
-        return [
-            [
-                "~name:'\"awe'\''some\"'",
-                'cool-options',
-                function (AbstractInvokable $invokable = null) {
-                    return new Options('cool-options', $this->getCoolOptionsBuilder(), $invokable);
-                },
-                function () {
-                    return ['name' => "\"awe'some\""];
-                }
-            ],
-            [
-                "'\"awe'\''some\"'",
-                'args',
-                function (AbstractInvokable $invokable = null) {
-                    return new Arguments($invokable);
-                },
-                function () {
-                    return ['name' => "\"awe'some\""];
-                }
-            ],
-            [
-                "--name='\"awe'\''some\"'",
-                'long-options',
-                function (AbstractInvokable $invokable = null) {
-                    return new LongOptions($invokable);
-                },
-                function () {
-                    return ['name' => "\"awe'some\""];
-                }
-            ],
-            [
-                "-n '\"awe'\''some\"'",
-                'short-options',
-                function (AbstractInvokable $invokable = null) {
-                    return new ShortOptions($invokable);
-                },
-                function () {
-                    return ['n' => "\"awe'some\""];
-                }
-            ],
-        ];
+        return require __DIR__ . '/OptionsTest.string.dataset.php';
     }
 
     /**
@@ -208,48 +167,7 @@ class OptionsTest extends PHPUnit_Framework_TestCase
      */
     public function provideOptionConfigsForArrayTest()
     {
-        return [
-            [
-                "~username:'light' ~host:'localhost' ~force ~name:'\"awe'\''some\"'",
-                'cool-options',
-                function (AbstractInvokable $invokable = null) {
-                    return new Options('cool-options', $this->getCoolOptionsBuilder(), $invokable);
-                },
-                function () {
-                    return ['username' => 'light', 'host' => 'localhost', 'force' => null, 'name' => "\"awe'some\""];
-                }
-            ],
-            [
-                "'light' 'localhost'  '\"awe'\''some\"'",
-                'args',
-                function (AbstractInvokable $invokable = null) {
-                    return new Arguments($invokable);
-                },
-                function () {
-                    return ['username' => 'light', 'host' => 'localhost', 'force' => null, 'name' => "\"awe'some\""];
-                }
-            ],
-            [
-                "--username='light' --host='localhost' --force --name='\"awe'\''some\"'",
-                'long-options',
-                function (AbstractInvokable $invokable = null) {
-                    return new LongOptions($invokable);
-                },
-                function () {
-                    return ['username' => 'light', 'host' => 'localhost', 'force' => null, 'name' => "\"awe'some\""];
-                }
-            ],
-            [
-                "-U 'light' -h 'localhost' -f -n '\"awe'\''some\"'",
-                'short-options',
-                function (AbstractInvokable $invokable = null) {
-                    return new ShortOptions($invokable);
-                },
-                function () {
-                    return ['U' => 'light', 'h' => 'localhost', 'f' => null, 'n' => "\"awe'some\""];
-                }
-            ],
-        ];
+        return require __DIR__ . '/OptionsTest.array.dataset.php';
     }
 
     /**
